\name{fastHorseshoe-package}
\alias{fastHorseshoe-package}
\docType{package}
\title{
  Fast Bayesian horseshoe regression implemented by the elliptical slice sampler.
}
\description{
  The elliptical slice sampler for Bayesian linear regression with shrinakge priors such as horseshoe, double-exponential prior. The main function can also implement user-specified prior functions. 
}
\details{
  For an overview of the package including list of functions and demos, please run
  
  \code{help(package="fastHorseshoe")}
}
\author{
P. Richard Hahn, Jingyu He and Hedibert Lopes

Maintainer: Jingyu He \email{jingyu.he@chicagobooth.edu}
}
\references{
Hahn, P. Richard, Jingyu He, and Hedibert Lopes. \emph{Elliptical slice sampling for Bayesian shrinkage regression with applications to causal inference} (2016).
}
\keyword{ package }
\seealso{
  fastHorseshoe
}
